added sr_latch test
