+++ Sim/tests/common/Makefile
@@ -1,6 +1,6 @@
 TOPLEVEL_LANG = verilog
 
-VERILOG_SOURCES = ../../hdl/common/d_latch.v ../../hdl/common/d_flip_flop.v ../../hdl/common/dm74ls283_quad_adder.v ../../hdl/common/sn54173_quad_flip_flop.v ../../hdl/common/full_adder.v ./test_module.v
+VERILOG_SOURCES = ../../hdl/common/d_latch.v ../../hdl/common/d_flip_flop.v ../../hdl/common/dm74ls283_quad_adder.v ../../hdl/common/sn54173_quad_flip_flop.v ../../hdl/common/full_adder.v ../../hdl/common/sr_latch.v ./test_module.v
 
 TOPLEVEL = test_module
